Do not export private components, rename DropdownItem -> DropdownItemCore
