+++ src/components/Combobox/AutoCompleteCombobox.js
@@ -2,7 +2,9 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import isEmpty from 'lodash-es/isEmpty';
 import isObject from 'lodash-es/isObject';
-import { ComboboxCore, BaseDropdownItem, ComboboxListbox } from './components';
+import ComboboxCore from './components/ComboboxCore';
+import ComboboxListbox from './components/ComboboxListbox';
+import { BaseDropdownItem } from './components/DropdownItems';
 import { InputRaw } from '../Form';
 import { makeInputAddHandler } from './utils/helpers';
 import { itemType } from './utils/constants';
