+++ src/components/Combobox/components/DropdownItems/HeaderDropdownItem.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { DropdownItem } from './DropdownItem';
+import { DropdownItemCore } from './DropdownItemCore';
 import { preventDefault } from '../../utils/helpers';
 
 export const HeaderDropdownItem = ({ id, label }) => (
-  <DropdownItem
+  <DropdownItemCore
     className="slds-listbox__option_plain"
     id={id}
     isHeader
     onSelect={preventDefault}
   >
     <h3 className="slds-text-title_caps">{label}</h3>
-  </DropdownItem>
+  </DropdownItemCore>
 );
 
 HeaderDropdownItem.propTypes = {
