+++ src/components/Combobox/components/index.js
@@ -1,13 +1,4 @@
-import ComboboxCore from './ComboboxCore';
-import ComboboxDropdown from './ComboboxDropdown';
-import ComboboxDropdownLists from './ComboboxDropdownLists';
 import ComboboxListbox from './ComboboxListbox';
 
 export * from './DropdownItems';
-
-export {
-  ComboboxCore,
-  ComboboxDropdown,
-  ComboboxDropdownLists,
-  ComboboxListbox,
-};
+export { ComboboxListbox };
